+++ Framework/Audio/Synthesizer/Oscillators/SquareWaveOscillator.cs
@@ -9,7 +9,7 @@
     public sealed class SquareWaveOscillator : IOscillator {
 
         /// <inheritdoc/>
-        public double GetSignal(double time, double frequency, double volume) {
+        public float GetSignal(float time, float frequency, float volume) {
             return Math.Sin(frequency * time * 2D * Math.PI) >= 0D ? volume : -volume;
         }
     }
